backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
pydantic==2.5.2
pydantic[email]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
google-generativeai
PyPDF2==3.0.1
python-docx==1.0.1
